requirements.txt
asgiref==3.8.1
Django==5.0.7
django-filter==24.2
django-phonenumber-field==8.0.0
django-seed==0.3.1
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==26.0.0
inflection==0.5.1
Markdown==3.6
mysqlclient==2.2.4
packaging==24.1
phonenumbers==8.13.40
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
toposort==1.10
tzdata==2024.1
uritemplate==4.1.1
